+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [0.9.7] 2020
 ### Improvements
 - AppImage
-    - Manual file installation:
+    - Manual file installation/upgrade:
         - default search path set to '~/Downloads'
         - trying to auto-fill the 'Name' and 'Version' fields